Move app.js to src for easier make a build
