+++ src/app.js
@@ -2,11 +2,10 @@ var http = require('http');
 var port = process.env.PORT || 2345;
 
 module.exports = http.createServer(function (req, res) {
-  res.writeHead(404, { 'Content-Type': 'text/plain' });
-
   if (req.url.indexOf('/suggestions') === 0) {
+    res.writeHead(404, { 'Content-Type': 'json/plain' });
     res.end(JSON.stringify({
-      suggestions: [],
+      suggestions: []
     }));
   } else {
     res.end();
